Arztverwaltung/server/requirements.txt
certifi>=2020.12.5
cffi>=1.14.5
chardet>=4.0.0
click>=7.1.2
colorama>=0.4.4
cors>=1.0.1
cryptography>=3.4.7
filelock>=3.0.12
Flask>=1.1.2
Flask-Cors>=3.0.10
Flask-MySQL>=1.5.2
future>=0.18.2
gevent>=21.1.2
greenlet>=1.1.0
idna>=2.10
itsdangerous>=1.1.0
Jinja2>=2.11.3
MarkupSafe>=1.1.1
mysql-connector-python>=8.0.24
mysqlclient>=2.0.3
protobuf>=3.16.0
pycparser>=2.20
PyMySQL>=1.0.2
PySocks>=1.7.1
requests>=2.25.1
requests-file>=1.5.1
six>=1.16.0
tldextract>=3.1.0
urllib3>=1.26.4
Werkzeug>=1.0.1
zope.event>=4.5.0
zope.interface>=5.4.0
